Fix link space prefixes in head
The .html extension combined with liquid syntax was tricking vs code
into thinking there had to be a space after a double quote. This was
adding a `%20`/space before the asset name in the url and could not be
found as a result.

All content is properly loading now, in addition to the newly added
favicons!